update bt scripts

+++ bt-install.sh
@@ -3,9 +3,12 @@ PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin
 export PATH
 LANG=en_US.UTF-8
 
-clear
-echo -e "\033[31m若在使用中发现问题，请及时反馈！ \033[0m"
-sleep 3s
+CURL_CHECK=$(which curl)
+if [ "$?" == "0" ];then
+	curl -sS --connect-timeout 10 -m 10 https://www.bt.cn/api/wpanel/SetupCount > /dev/null 2>&1
+else
+	wget -O /dev/null -o /dev/null -T 5 https://www.bt.cn/api/wpanel/SetupCount
+fi
 
 if [ $(whoami) != "root" ];then
 	echo "请使用root权限执行宝塔安装命令！"
@@ -134,7 +137,30 @@ Service_Add(){
 		update-rc.d bt defaults
 	fi 
 }
-
+Set_Centos_Repo(){
+	HUAWEI_CHECK=$(cat /etc/motd |grep "Huawei Cloud")
+	if [ "${HUAWEI_CHECK}" ] && [ "${is64bit}" == "64" ];then
+		\cp -rpa /etc/yum.repos.d/ /etc/yumBak
+		sed -i 's/mirrorlist/#mirrorlist/g' /etc/yum.repos.d/CentOS-*.repo
+		sed -i 's|#baseurl=http://mirror.centos.org|baseurl=http://vault.epel.cloud|g' /etc/yum.repos.d/CentOS-*.repo
+		rm -f /etc/yum.repos.d/epel.repo
+		rm -f /etc/yum.repos.d/epel-*
+	fi
+	ALIYUN_CHECK=$(cat /etc/motd|grep "Alibaba Cloud ")
+	if [  "${ALIYUN_CHECK}" ] && [ "${is64bit}" == "64" ] && [ ! -f "/etc/yum.repos.d/Centos-vault-8.5.2111.repo" ];then
+		rename '.repo' '.repo.bak' /etc/yum.repos.d/*.repo
+		wget https://mirrors.aliyun.com/repo/Centos-vault-8.5.2111.repo -O /etc/yum.repos.d/Centos-vault-8.5.2111.repo
+		wget https://mirrors.aliyun.com/repo/epel-archive-8.repo -O /etc/yum.repos.d/epel-archive-8.repo
+		sed -i 's/mirrors.cloud.aliyuncs.com/url_tmp/g'  /etc/yum.repos.d/Centos-vault-8.5.2111.repo &&  sed -i 's/mirrors.aliyun.com/mirrors.cloud.aliyuncs.com/g' /etc/yum.repos.d/Centos-vault-8.5.2111.repo && sed -i 's/url_tmp/mirrors.aliyun.com/g' /etc/yum.repos.d/Centos-vault-8.5.2111.repo
+		sed -i 's/mirrors.aliyun.com/mirrors.cloud.aliyuncs.com/g' /etc/yum.repos.d/epel-archive-8.repo
+	fi
+	MIRROR_CHECK=$(cat /etc/yum.repos.d/CentOS-Linux-AppStream.repo |grep "[^#]mirror.centos.org")
+	if [ "${MIRROR_CHECK}" ] && [ "${is64bit}" == "64" ];then
+		\cp -rpa /etc/yum.repos.d/ /etc/yumBak
+		sed -i 's/mirrorlist/#mirrorlist/g' /etc/yum.repos.d/CentOS-*.repo
+		sed -i 's|#baseurl=http://mirror.centos.org|baseurl=http://vault.epel.cloud|g' /etc/yum.repos.d/CentOS-*.repo
+	fi
+}
 get_node_url(){
 	if [ ! -f /bin/curl ];then
 		if [ "${PM}" = "yum" ]; then
@@ -143,10 +169,17 @@ get_node_url(){
 			apt-get install curl -y
 		fi
 	fi
+
+	if [ -f "/www/node.pl" ];then
+		download_Url=$(cat /www/node.pl)
+		echo "Download node: $download_Url";
+		echo '---------------------------------------------';
+		return
+	fi
 	
 	echo '---------------------------------------------';
 	echo "Selected download node...";
-	nodes=(http://dg2.bt.cn http://dg1.bt.cn http://125.90.93.52:5880 http://36.133.1.8:5880 http://123.129.198.197 http://38.34.185.130 http://103.224.251.67:5880 http://128.1.164.196);
+	nodes=(http://dg2.bt.cn http://dg1.bt.cn http://125.90.93.52:5880 http://36.133.1.8:5880 http://123.129.198.197 http://38.34.185.130 http://116.213.43.206:5880 http://128.1.164.196);
 	tmp_file1=/dev/shm/net_test1.pl
 	tmp_file2=/dev/shm/net_test2.pl
 	[ -f "${tmp_file1}" ] && rm -f ${tmp_file1}
@@ -189,7 +222,6 @@ get_node_url(){
 	rm -f $tmp_file1
 	rm -f $tmp_file2
 	download_Url=$NODE_URL
-	downloads_Url=http://download.moetas.com/ltd
 	echo "Download node: $download_Url";
 	echo '---------------------------------------------';
 }
@@ -210,6 +242,9 @@ Remove_Package(){
 Install_RPM_Pack(){
 	yumPath=/etc/yum.conf
 	Centos8Check=$(cat /etc/redhat-release | grep ' 8.' | grep -iE 'centos|Red Hat')
+	if [ "${Centos8Check}" ];then
+		Set_Centos_Repo
+	fi	
 	isExc=$(cat $yumPath|grep httpd)
 	if [ "$isExc" = "" ];then
 		echo "exclude=httpd nginx php mysql mairadb python-psutil python2-psutil" >> $yumPath
@@ -222,9 +257,9 @@ Install_RPM_Pack(){
 	#	curl -Ss --connect-timeout 3 -m 60 http://download.bt.cn/install/yumRepo_select.sh|bash
 	#fi
 	
-	#尝试同步时间(从www.moetas.com)
+	#尝试同步时间(从bt.cn)
 	echo 'Synchronizing system time...'
-	getBtTime=$(curl -sS --connect-timeout 3 -m 60 http://www.moetas.com/api/index/get_time)
+	getBtTime=$(curl -sS --connect-timeout 3 -m 60 http://www.bt.cn/api/index/get_time)
 	if [ "${getBtTime}" ];then	
 		date -s "$(date -d @$getBtTime +"%Y-%m-%d %H:%M:%S")"
 	fi
@@ -279,6 +314,12 @@ Install_Deb_Pack(){
 	#echo 'Synchronizing system time...'
 	#ntpdate 0.asia.pool.ntp.org
 	#apt-get upgrade -y
+	LIBCURL_VER=$(dpkg -l|grep libcurl4|awk '{print $3}')
+	if [ "${LIBCURL_VER}" == "7.68.0-1ubuntu2.8" ];then
+		apt-get remove libcurl4 -y
+		apt-get install curl -y
+	fi
+
 	debPacks="wget curl libcurl4-openssl-dev gcc make zip unzip tar openssl libssl-dev gcc libxml2 libxml2-dev zlib1g zlib1g-dev libjpeg-dev libpng-dev lsof libpcre3 libpcre3-dev cron net-tools swig build-essential libffi-dev libbz2-dev libncurses-dev libsqlite3-dev libreadline-dev tk-dev libgdbm-dev libdb-dev libdb++-dev libpcap-dev xz-utils git";
 	apt-get install -y $debPacks --force-yes
 
@@ -286,7 +327,7 @@ Install_Deb_Pack(){
 	do
 		packCheck=$(dpkg -l ${debPack})
 		if [ "$?" -ne "0" ] ;then
-			apt-get install -y debPack
+			apt-get install -y $debPack
 		fi
 	done
 
@@ -311,6 +352,9 @@ Get_Versions(){
 		os_version=$(cat $redhat_version_file|grep CentOS|grep -Eo '([0-9]+\.)+[0-9]+'|grep -Eo '^[0-9]')
 		if [ "${os_version}" = "5" ];then
 			os_version=""
+		fi
+		if [ -z "${os_version}" ];then
+			os_version=$(cat /etc/redhat-release |grep Stream|grep -oE 8)
 		fi
 	else
 		os_type='ubuntu'
@@ -337,7 +381,15 @@ Get_Versions(){
 			if [ "$os_version" = "19" ];then
 				os_version=""
 			fi
-
+			if [ "$os_version" = "21" ];then
+				os_version=""
+			fi
+			if [ "$os_version" = "20" ];then
+				os_version2004=$(cat /etc/issue|grep 20.04)
+				if [ -z "${os_version2004}" ];then
+					os_version=""
+				fi
+			fi
 		fi
 	fi
 }
@@ -357,10 +409,50 @@ Install_Python_Lib(){
 				$pyenv_path/pyenv/bin/pip install -r $pyenv_path/pyenv/pip.txt
 			fi
 			source $pyenv_path/pyenv/bin/activate
+			chmod -R 700 $pyenv_path/pyenv/bin
 			return
 		else
 			rm -rf $pyenv_path/pyenv
 		fi
+	fi
+
+	is_loongarch64=$(uname -a|grep loongarch64)
+	if [ "$is_loongarch64" != "" ] && [ -f "/usr/bin/yum" ];then
+		yumPacks="python3-devel python3-pip python3-psutil python3-gevent python3-pyOpenSSL python3-paramiko python3-flask python3-rsa python3-requests python3-six python3-websocket-client"
+		yum install -y ${yumPacks}
+		for yumPack in ${yumPacks}
+		do
+			rpmPack=$(rpm -q ${yumPack})
+			packCheck=$(echo ${rpmPack}|grep not)
+			if [ "${packCheck}" ]; then
+				yum install ${yumPack} -y
+			fi
+		done
+
+		pip3 install -U pip
+		pip3 install Pillow psutil pyinotify pycryptodome upyun oss2 pymysql qrcode qiniu redis pymongo Cython configparser cos-python-sdk-v5 supervisor gevent-websocket pyopenssl
+		pip3 install flask==1.1.4
+		pip3 install Pillow -U
+
+		pyenv_bin=/www/server/panel/pyenv/bin
+		mkdir -p $pyenv_bin
+		ln -sf /usr/local/bin/pip3 $pyenv_bin/pip
+		ln -sf /usr/local/bin/pip3 $pyenv_bin/pip3
+		ln -sf /usr/local/bin/pip3 $pyenv_bin/pip3.7
+
+		if [ -f "/usr/bin/python3.7" ];then
+			ln -sf /usr/bin/python3.7 $pyenv_bin/python
+			ln -sf /usr/bin/python3.7 $pyenv_bin/python3
+			ln -sf /usr/bin/python3.7 $pyenv_bin/python3.7
+		elif [ -f "/usr/bin/python3.6"  ]; then
+			ln -sf /usr/bin/python3.6 $pyenv_bin/python
+			ln -sf /usr/bin/python3.6 $pyenv_bin/python3
+			ln -sf /usr/bin/python3.6 $pyenv_bin/python3.7
+		fi
+
+		echo > $pyenv_bin/activate
+
+		return
 	fi
 
 	py_version="3.7.8"
@@ -470,14 +562,18 @@ Install_Bt(){
 	mkdir -p /www/backup/database
 	mkdir -p /www/backup/site
 
+	if [ ! -d "/etc/init.d" ];then
+		mkdir -p /etc/init.d
+	fi
+
 	if [ -f "/etc/init.d/bt" ]; then
 		/etc/init.d/bt stop
 		sleep 1
 	fi
 
-	wget -O /etc/init.d/bt ${downloads_Url}/install/src/bt6.init -T 10
+	wget -O /etc/init.d/bt ${download_Url}/install/src/bt6.init -T 10
 	wget -O /www/server/panel/install/public.sh ${download_Url}/install/public.sh -T 10
-	wget -O panel.zip ${downloads_Url}/install/src/panel6.zip -T 10
+	wget -O panel.zip ${download_Url}/install/src/panel6.zip -T 10
 
 	if [ -f "${setup_path}/server/panel/data/default.db" ];then
 		if [ -d "/${setup_path}/server/panel/old_data" ];then
@@ -496,6 +592,7 @@ Install_Bt(){
 		if [ "${PM}" = "yum" ]; then
 			yum install unzip -y
 		elif [ "${PM}" = "apt-get" ]; then
+			apt-get update
 			apt-get install unzip -y
 		fi
 	fi
@@ -526,11 +623,18 @@ Install_Bt(){
 	chmod -R +x ${setup_path}/server/panel/script
 	ln -sf /etc/init.d/bt /usr/bin/bt
 	echo "${panelPort}" > ${setup_path}/server/panel/data/port.pl
-	wget -O /etc/init.d/bt ${downloads_Url}/install/src/bt7.init -T 10
-	wget -O /www/server/panel/init.sh ${downloads_Url}/install/src/bt7.init -T 10
-	sed -i 's/[0-9\.]\+[ ]\+www.bt.cn//g' /etc/hosts
+	wget -O /etc/init.d/bt ${download_Url}/install/src/bt7.init -T 10
+	wget -O /www/server/panel/init.sh ${download_Url}/install/src/bt7.init -T 10
+	wget -O /www/server/panel/data/softList.conf ${download_Url}/install/conf/softList.conf
 }
 Set_Bt_Panel(){
+	Run_User="www"
+	wwwUser=$(cat /etc/passwd|cut -d ":" -f 1|grep ^www$)
+	if [ "${wwwUser}" != "www" ];then
+		groupadd ${Run_User}
+		useradd -s /sbin/nologin -g ${Run_User} ${Run_User}
+	fi
+
 	password=$(cat /dev/urandom | head -n 16 | md5sum | head -c 8)
 	sleep 1
 	admin_auth="/www/server/panel/data/admin_path.pl"
@@ -538,6 +642,9 @@ Set_Bt_Panel(){
 		auth_path=$(cat /dev/urandom | head -n 16 | md5sum | head -c 8)
 		echo "/${auth_path}" > ${admin_auth}
 	fi
+	chmod -R 700 $pyenv_path/pyenv/bin
+	/www/server/panel/pyenv/bin/pip3 install flask -U
+	/www/server/panel/pyenv/bin/pip3 install flask-sock
 	auth_path=$(cat ${admin_auth})
 	cd ${setup_path}/server/panel/
 	/etc/init.d/bt start
@@ -621,13 +728,13 @@ Set_Firewall(){
 }
 Get_Ip_Address(){
 	getIpAddress=""
-	getIpAddress=$(curl -sS --connect-timeout 10 -m 60 https://www.moetas.com/Api/getIpAddress)
+	getIpAddress=$(curl -sS --connect-timeout 10 -m 60 https://www.bt.cn/Api/getIpAddress)
 	if [ -z "${getIpAddress}" ] || [ "${getIpAddress}" = "0.0.0.0" ]; then
 		isHosts=$(cat /etc/hosts|grep 'www.bt.cn')
 		if [ -z "${isHosts}" ];then
 			echo "" >> /etc/hosts
-			echo "103.224.251.67 www.bt.cn" >> /etc/hosts
-			getIpAddress=$(curl -sS --connect-timeout 10 -m 60 https://www.moetas.com/Api/getIpAddress)
+			echo "116.213.43.206 www.bt.cn" >> /etc/hosts
+			getIpAddress=$(curl -sS --connect-timeout 10 -m 60 https://www.bt.cn/Api/getIpAddress)
 			if [ -z "${getIpAddress}" ];then
 				sed -i "/bt.cn/d" /etc/hosts
 			fi
@@ -726,6 +833,6 @@ echo -e "=================================================================="
 endTime=`date +%s`
 ((outTime=($endTime-$startTime)/60))
 echo -e "Time consumed:\033[32m $outTime \033[0mMinute!"
-echo -e "\033[31m已经安装完毕，欢迎使用！ \033[0m"  
-rm -rf install.sh
+
+
